% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agreements.R
\name{agreement}
\alias{agreement}
\title{Agreement}
\usage{
agreement(table)
}
\arguments{
\item{table}{ a data matrix or table with equal number of columns and rows.}
}
\value{
An S3 object containing the proportion of overall agreement.
}
\description{
The proportion of overall agreement between 2 or more raters is calculated by the sum over the diagonal of the cross table divided by the total sum of the cells.
}
\examples{
df <- data.frame(r1=factor(c(1,0,1,0,0,1,1,0,0,0,1,1,0,1,1)),
                 r2=factor(c(1,1,1,1,0,1,1,0,0,0,1,1,0,1,0)),
                 r3=factor(c(1,1,1,0,0,0,1,1,1,0,0,1,0,1,1)),
                 r4=factor(c(1,1,1,1,1,1,1,1,1,1,1,1,1,1,1)))
table <- sumtable(df=df, ratings=c("r1", "r2", "r3", "r4"), levels=c("0","1"))
agreement(table)

df <- data.frame(r1=factor(c(1,2,2,0,3,3,1,0,3,0,2,2,0,3,1)),
                 r2=factor(c(1,1,1,0,3,3,1,0,1,0,2,2,0,2,1)),
                 r3=factor(c(1,1,1,3,3,2,1,0,1,0,2,2,0,3,1)),
                 r4=factor(c(1,2,1,0,3,3,1,0,3,0,2,2,0,2,1)))
table <- sumtable(df=df, ratings=c("r1", "r2", "r3", "r4"), levels=c("0","1", "2", "3"))
agreement(table)
}

