% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIagreement.R
\name{CIagreement}
\alias{CIagreement}
\title{Confidence interval around the Agreement}
\usage{
CIagreement(p, n, m, level = 0.95, correction = "continuity")
}
\arguments{
\item{p}{Proportion; a numeric value between 0 and 1}

\item{n}{The sample size of the sample that is rated}

\item{m}{The number of raters}

\item{level}{Confidence level; default is 0.95.}

\item{correction}{A character value indicating the method used for correction of the interval, options are "continuity" or "Fleis".}
}
\value{
A vector giving the lower and upper confidence limit around the probability.
}
\description{
The confidence interval estimated around the proportion of agreement between 2 or more raters.
}
\examples{
df <- data.frame(r1=factor(c(1,0,1,0,0,1,1,0,0,0,1,1,0,1,1)),
                 r2=factor(c(1,1,1,1,0,1,1,0,0,0,1,1,0,1,0)),
                 r3=factor(c(1,1,1,0,0,0,1,1,1,0,0,1,0,1,1)),
                 r4=factor(c(1,1,1,1,1,1,1,1,1,1,1,1,1,1,1)))
table <- sumtable(df=df, ratings=c("r1", "r2", "r3", "r4"), levels=c("0","1"))
p <- agreement(table)
CIagreement(p=p, n=nrow(df), m=ncol(df))

}
