% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old.R
\name{CIbootagreement}
\alias{CIbootagreement}
\title{Bootstrapped confidence interval around the Agreement}
\usage{
CIbootagreement(data, cat1 = NULL, cat2 = NULL, b = 1000, alpha = 0.05)
}
\arguments{
\item{data}{data.frame or table}

\item{cat1}{A character indicating the category for which specific agreement should be obtained.}

\item{cat2}{A character indicating the category to which the specific agreement should be compared, if left empty all other categories are used.}

\item{b}{Number of bootstrap iterations.}

\item{alpha}{Confidence level; default is 0.95.}
}
\value{
A vector giving the lower and upper confidence limit around the probability.
}
\description{
The confidence interval estimated around the proportion of agreement between 2 or more raters.
}
\examples{
df <- data.frame(r1=factor(c(1,0,1,0,0,1,1,0,0,0,1,1,0,1,1)),
                 r2=factor(c(1,1,1,1,0,1,1,0,0,0,1,1,0,1,0)),
                 r3=factor(c(1,1,1,0,0,0,1,1,1,0,0,1,0,1,1)),
                 r4=factor(c(1,1,1,1,1,1,1,1,1,1,1,1,1,1,1)))
CIbootagreement(data=df)


}
