% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{diagnoses}
\alias{diagnoses}
\title{diagnoses}
\format{
A data frame with 30 rows and 6 variables:
\describe{
  \item{rater5}{diagnoses of rater 5}
  \item{rater1}{diagnoses of rater 1}
  \item{rater3}{diagnoses of rater 3}
  \item{rater2}{diagnoses of rater 2}
  \item{rater6}{diagnoses of rater 6}
  \item{rater4}{diagnoses of rater 4}
}
}
\source{
data-raw/R/diagnoses.R
}
\usage{
diagnoses
}
\description{
For the nominal data example we use a data set that was used in a paper by
Fleis (1971). In this data patients are diagnosed in 5 categories:
Depression, Personality Disorder, Schizophrenia, Neurosis, and Other by six
raters.
}
\keyword{datasets}
