% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma-plot.R
\name{geom_ba}
\alias{geom_ba}
\title{Bland-Altman Agreement Plot}
\usage{
geom_ba(
  mapping = NULL,
  data = NULL,
  stat = "BA",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by aes(). If specified and inherit.aes = TRUE (the default), it is combined with the default mapping at the top level of the plot. You must supply mapping if there is no plot mapping.}

\item{data}{The data to be displayed in this layer}

\item{stat}{Generally unused}

\item{position}{Generally unused}

\item{na.rm}{If FALSE, the default, missing values are removed with a warning. If TRUE, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends? NA, the default, includes if any aesthetics are mapped. FALSE never includes, and TRUE always includes. It can also be a named logical vector to finely select the aesthetics to display.}

\item{inherit.aes}{If FALSE, overrides the default aesthetics, rather than combining with them.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{ggplot2::layer()}}.}
}
\description{
Bland-Altman Agreement Plot
}
\examples{
d <- data.frame(x = rnorm(100), y = rnorm(100)) |>
  dplyr::mutate(
    .difference = x - y,
    .average = (x + y) / 2
  )
d |>
  ggplot2::ggplot(ggplot2::aes(x = .average, y = .difference)) +
  geom_ba()

d$group <- rep(c(1, 2), each = 50)

d |>
  ggplot2::ggplot(ggplot2::aes(x = .average, y = .difference)) +
  ggplot2::facet_wrap(~group) +
  geom_ba()

}
