% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icc_agreement.R
\name{icc_agreement}
\alias{icc_agreement}
\title{ICC agreement}
\usage{
icc_agreement(model, alpha = 0.05)
}
\arguments{
\item{model}{merMod object result from \code{icc_model()}.}

\item{alpha}{confidence interval level, default \code{alpha = 0.05}.}
}
\value{

}
\description{
ICC agreement
}
