% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icc_model.R
\name{icc_model}
\alias{icc_model}
\title{ICC model}
\usage{
icc_model(data, cols = colnames(data))
}
\arguments{
\item{data}{data.frame with observerd in the columns and subjects in the rows}

\item{cols}{character vector with the column names to be used as observers.}
}
\value{

}
\description{
Calculate model for the ICC estimations. The basic model contains a random intercept for the id-level and a random intercept for the observer level. For the ICC calculations, different variance and error components are used.
}
