\name{judgeRatings}
\docType{data}
\alias{judgeRatings}
\title{Ratings of Different Judges}
\description{
  The ratings of judges on a specific characteristic.
}
\usage{judgeRatings}
\format{
  A matrix presenting the ratings of four judges on six people.
}
\source{
  B. J. Winer (1971) \emph{Statistical principles in experimental design},
  (2nd ed.). McGraw-Hill, New York
 }
\keyword{datasets}
