% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kappa.R
\name{kappa}
\alias{kappa}
\title{Kappa for agreement with multiple raters}
\usage{
kappa(data, confint = FALSE, alpha = 0.05, k = ncol(data), n = nrow(data), ...)
}
\arguments{
\item{data}{A data frame or table with equal number of columns and rows. Or a data frame that contains the scores for each rater in each column.}

\item{confint}{Logical indicator for confidence interval}

\item{alpha}{Confidence level, default = 0.05.}

\item{k}{number of raters; default \code{k = ncol(data)}}

\item{n}{sample size; default \code{n = nrow(data)}}

\item{\dots}{options for sumtable if \code{data = data.frame}}
}
\value{

}
\description{
Kappa for agreement with multiple raters
}
\examples{
df <- data.frame(r1=factor(c(1,0,1,0,0,1,1,0,0,0,1,1,0,1,1)),
                 r2=factor(c(1,1,1,1,0,1,1,0,0,0,1,1,0,1,0)),
                 r3=factor(c(1,1,1,0,0,0,1,1,1,0,0,1,0,1,1)),
                 r4=factor(c(1,1,1,1,1,1,1,1,1,1,1,1,1,1,1)))
table <- sumtable(df=df, ratings=c("r1", "r2", "r3", "r4"), levels=c("0","1"))
kappa(df)
kappa(table)

df <- data.frame(r1=factor(c(1,2,2,0,3,3,1,0,3,0,2,2,0,3,1)),
                 r2=factor(c(1,1,1,0,3,3,1,0,1,0,2,2,0,2,1)),
                 r3=factor(c(1,1,1,3,3,2,1,0,1,0,2,2,0,3,1)),
                 r4=factor(c(1,2,1,0,3,3,1,0,3,0,2,2,0,2,1)))
table <- sumtable(df=df, ratings=c("r1", "r2", "r3", "r4"), levels=c("0","1", "2", "3"))
kappa(df)
kappa(table)
kappa(df, confint = TRUE)
kappa(table)
}
