\name{lesionBurden}
\docType{data}
\alias{lesionBurden}
\alias{lesionBurden.M}
\alias{lesionBurden.G}
\title{Total Lesion Burden}
\description{
  The total lesion volumes measured manually and by an automated
  technique known as Geometrically Constrained Region Growth.
  
}
\usage{lesionBurden}
\format{
  lesionBurden.M is a matrix presenting the manually measured volumes 
  on three patients each with ten replicates.
  
  lesionBurden.G is a matrix presenting the automatically measured volumes 
  on three patients each with ten replicates.
}
\source{
  Mohamed M Shoukri, Nasser Elkum and Stephen D Walter (2006)
  Interval estimation and optimal design for the within-subject
  coefficient of variation for continuous and binary variables.
  \emph{BMC Medical Research Methodology}
  \bold{6} 24
 }
\keyword{datasets}
