% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_icc.R
\name{n_icc}
\alias{n_icc}
\title{Sample size for ICC}
\usage{
n_icc(beta = 0.8, alpha = 0.05, k = 3, icc, icc_lower)
}
\arguments{
\item{beta}{power level, default \code{beta = 0.8}}

\item{alpha}{confidence level, default \code{alpha = 0.05}}

\item{k}{number of raters, default \code{k = 3}}

\item{icc}{reliability coefficient}

\item{icc_lower}{lower limit of the reliability coefficient icc}
}
\value{

}
\description{
Sample size for ICC
}
