\name{petVT}
\docType{data}
\alias{petVT}
\alias{amygdala}
\alias{cerebellum}
\alias{brainStem}
\title{PET Total Volume of Distribution}
\description{
  Test/retest data for total volume of distribution (VT) from three published PET studies.
}
\usage{petVT}
\format{
  A list presenting the VT from three studies. The first component is
  the data from Table 6 of  Ogden et al. (2007). The second component is
  the data from Table 3 of Hostetler et al. (2013). The third component is
  the data from Table II of Gunn et al. (2011).
}
\source{
  R Todd Ogden et al. (2007)
  In vivo quantification of serotonin transporters using [11C]DASB
  and positron emission tomography in humans: modeling considerations
  \emph{Journal of Cerbral Blood Flow & Metabolism} \bold{27} 205-217
  
  Eric D. Hostetler et al. (2013)
  In vivo quantification of calcitonin gene-related peptide receptor occupancy by telcagepant in rhesus monkey and
  human brain using the positron emission tomography tracer [11C]MK-4232
  \emph{The Journal of Pharmacology and Experimental Therapeutics} \bold{347} 478-486

  Roger N. Gunn et al. (2011)
  Translational characterization of [11C]GSK931145, a PET ligand for the
  Glycine transporter type 1
  \emph{SYNAPSE} \bold{65} 1319-1332

}
\keyword{datasets}
