% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old.R
\name{positive.agreement-deprecated}
\alias{positive.agreement-deprecated}
\title{positive agreement-deprecated}
\usage{
positive.agreement(data, cat1)
}
\arguments{
\item{data}{A data matrix or table with equal number of columns and rows. Or a data frame that contains the scores for each rater in each column.}

\item{specific}{A character vector indicating whether the \code{"positive"} or \code{"negative"} agreements should be obtained.}

\item{\dots}{optoins for sumtable.}
}
\value{
An S3 object containing the proportion of positive (or negative) agreement.
}
\description{
A function to obtain the positive or negative agreement between 2 or more raters when categories are dichotomous.
}
\seealso{
\code{\link{Agree-deprecated}}
}
\keyword{internal}
