% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agreements.R
\name{positive.agreement}
\alias{positive.agreement}
\title{positive agreement}
\usage{
positive.agreement(data, specific = "positive", ratings = NULL,
  levels = NULL, offdiag = NULL)
}
\arguments{
\item{data}{A data matrix or table with equal number of columns and rows. Or a data frame that contains the scores for each rater in each column.}

\item{specific}{A character vector indicating whether the \code{"positive"} or \code{"negative"} agreements should be obtained.}
}
\value{
An S3 object containing the proportion of positive (or negative) agreement.
}
\description{
A function to obtain the positive or negative agreement between 2 or more raters when categories are dichotomous. INCLUDE FORMULAS.
}
\examples{
#' df <- data.frame(r1=factor(c(1,0,1,0,0,1,1,0,0,0,1,1,0,1,1)),
                 r2=factor(c(1,1,1,1,0,1,1,0,0,0,1,1,0,1,0)),
                 r3=factor(c(1,1,1,0,0,0,1,1,1,0,0,1,0,1,1)),
                 r4=factor(c(1,1,1,1,1,1,1,1,1,1,1,1,1,1,1)))
table <- sumtable(df=df, ratings=c("r1", "r2", "r3", "r4"), levels=c("0","1"))
positive.agreement(table, specific="positive")
}
