% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agreements.R
\name{specific.agreement}
\alias{specific.agreement}
\title{Specific agreement for more than 2 categories}
\usage{
specific.agreement(table, cat1, cat2 = NULL)
}
\arguments{
\item{table}{A data matrix or table with equal number of columns and rows.}

\item{cat1}{A character indicating the category for which specific agreement should be obtained.}

\item{cat2}{A character indicating the category to which the specific agreement should be compared, if left empty all other categories are used.}
}
\value{
An S3 object containing the proportion of specific agreement.
}
\description{
specific agreement when there are more than two categories (averages over discordant cells to correct for random rater combinations). INCLUDE FORMULAS.
Specific agreement for one category versus the others, for example very satisfied verus rest.
}
\examples{
#' df <- data.frame(r1=factor(c(1,2,2,0,3,3,1,0,3,0,2,2,0,3,1)),
                 r2=factor(c(1,1,1,0,3,3,1,0,1,0,2,2,0,2,1)),
                 r3=factor(c(1,1,1,3,3,2,1,0,1,0,2,2,0,3,1)),
                 r4=factor(c(1,2,1,0,3,3,1,0,3,0,2,2,0,2,1)))
table <- sumtable(df=df, ratings=c("r1", "r2", "r3", "r4"), levels=c("0","1", "2", "3"))
specific.agreement(table)
}
