% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agreements.R
\name{specific.agreement2}
\alias{specific.agreement2}
\title{Specific agreement for more than 2 categories}
\usage{
specific.agreement2(table)
}
\arguments{
\item{table}{A data matrix or table with equal number of columns and rows.}
}
\value{
betweentable The same as the input table, where the discordant cells are averaged to correct for random rater combinations.

specific.agreements

specific.agreements.prop
}
\description{
specific agreement when there are more than one categories (averages over discordant cells to correct for random rater combinations). INCLUDE FORMULAS.
}
\examples{
#' df <- data.frame(r1=factor(c(1,2,2,0,3,3,1,0,3,0,2,2,0,3,1)),
                 r2=factor(c(1,1,1,0,3,3,1,0,1,0,2,2,0,2,1)),
                 r3=factor(c(1,1,1,3,3,2,1,0,1,0,2,2,0,3,1)),
                 r4=factor(c(1,2,1,0,3,3,1,0,3,0,2,2,0,2,1)))
table <- sumtable(df=df, ratings=c("r1", "r2", "r3", "r4"), levels=c("0","1", "2", "3"))
specific.agreement2(table)
}

