% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumtable.R
\name{sumtable}
\alias{sumtable}
\title{Summed Table}
\usage{
sumtable(df, ratings = NULL, levels = NULL, offdiag = NULL)
}
\arguments{
\item{df}{The input data frame that contains the scores for each rater in each column}

\item{ratings}{A character vector that contains the names of the factor variables that need to be used as ratings}

\item{levels}{A character vector that contains the levels of the factors.}

\item{offdiag}{A logical parameter indicating if the of diagonal means should be used, default is TRUE with more than two raters}
}
\value{
Returns a contingency table, an object of class "table", an array of integer values.
}
\description{
This function can be used to calculate the sum of the tables for each combination of raters. The output table can be used to estimate the agreement statistics between two or more raters. First the crosstable is made for each combination of raters. This results in \code{2*(m-1)} tables, where \code{m} is the number of raters. Then these tables are summed to one table.
}
\examples{
df <- data.frame(r1=factor(c(1,0,1,0,0,1,1,0,0,0,1,1,0,1,1)),
                 r2=factor(c(1,1,1,1,0,1,1,0,0,0,1,1,0,1,0)),
                 r3=factor(c(1,1,1,0,0,0,1,1,1,0,0,1,0,1,1)),
                 r4=factor(c(1,1,1,1,1,1,1,1,1,1,1,1,1,1,1)))
sumtable(df=df, ratings=c("r1", "r2", "r3", "r4"), levels=c("0","1"))

df <- data.frame(r1=factor(c(1,2,2,0,3,3,1,0,3,0,2,2,0,3,1)),
                 r2=factor(c(1,1,1,0,3,3,1,0,1,0,2,2,0,2,1)),
                 r3=factor(c(1,1,1,3,3,2,1,0,1,0,2,2,0,3,1)),
                 r4=factor(c(1,2,1,0,3,3,1,0,3,0,2,2,0,2,1)))
sumtable(df=df, ratings=c("r1", "r2", "r3", "r4"), levels=c("0","1", "2", "3"))
}
