% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootCI.R
\name{bootCI}
\alias{bootCI}
\title{bootCI}
\usage{
bootCI(data, b = 1000, alpha = 0.05, fun, ...)
}
\arguments{
\item{data}{data.frame for which the bootstraps need to be performed to obtain
confidence intervals.}

\item{b}{number of bootstrap replications}

\item{alpha}{level of Confidence interval, default \code{alpha = 0.05}}

\item{fun}{function to estimate parameter for which the CI is determined.}

\item{\dots}{additional parameters for function}
}
\value{
named vector with confidence interval levels
}
\description{
bootCI
}
\examples{
agreement(diagnoses)
bootCI(data = diagnoses, fun = agreement)
agreement(diagnoses, specific = "4. Neurosis")
bootCI(data = diagnoses, fun = agreement, specific = "4. Neurosis")
}
