% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icc.R
\name{icc}
\alias{icc}
\title{Intra class correlation for rater reliability}
\usage{
icc(data, method = c("oneway", "agreement", "consistency"),
  cols = colnames(data), sem = TRUE, confint = FALSE, level = 0.95)
}
\arguments{
\item{data}{data.frame with repeated measures or observations in the columns and rated subjects in the rows}

\item{cols}{column names used for the repeated measures, default \code{cols = colnames(data)}}

\item{sem}{logical vector if sem should be returned}

\item{confint}{logical vector if confidence interval for icc should be computed (see details for formulas)}

\item{level}{the confidence level required.}

\item{icc}{icc method \code{icc = c("oneway", "agreement", "consistency")}, see details for explanatory formulas.}
}
\value{
matrix with relevant output
}
\description{
Intra class correlation for rater reliability
}
\examples{
icc <- 2
means <- c(1, rep(0,(5-1)))
means <- c(rep(0, 5))
means  <- means * 1
cov <- matrix(0.7,5,5)
diag(cov) <- 1 # large variance
cov <- cov*1 # adjust variance to small, medium or large
dat <- MASS::mvrnorm(means, cov, n=100)
dat <- as.data.frame(dat)
icc(dat)
}
