% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icc.R
\name{icc}
\alias{icc}
\title{Intra class correlation for rater reliability}
\usage{
icc(
  data,
  method = c("oneway", "agreement", "consistency"),
  cols = colnames(data),
  sem = TRUE,
  confint = TRUE,
  alpha = 0.05,
  var = FALSE,
  onemodel = FALSE
)
}
\arguments{
\item{data}{data.frame with repeated measures or observations in the columns
and rated subjects in the rows}

\item{method}{type of ICC that is returned, options are:
`c("oneway", "agreement", "consistency")`, the default returns all.}

\item{cols}{column names used for the repeated measures in the wide format,
default uses `cols = colnames(data)`}

\item{sem}{logical vector if standard error of measurement is returned.}

\item{confint}{logical vector if confidence interval for ICC is returned.}

\item{alpha}{the confidence level required, default `alpha = 0.05`.}

\item{var}{logical indicator if variance estimates are returned.}

\item{onemodel}{logical indicator if all ICC's should be computed from one
model to increase computation time, defalt `onemodel = FALSE`.}
}
\value{
matrix with relevant output
}
\description{
Intra class correlation for rater reliability
}
\examples{
mam <-
breast[,c("Mam1_score","Mam2_score", "Mam3_score")]
icc(data = mam, confint = TRUE, var = TRUE)
pch <- breast[,c("PCH1_score", "PCH2_score","PCH3_score",
"PCH4_score","PCH5_score")]
icc(data = pch)
icc(data = pch, confint = FALSE, var = TRUE)
}
\seealso{
[icc_agreement()] [icc_oneway()] [icc_consistency()]
}
