% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icc.R
\name{icc}
\alias{icc}
\title{Intra class correlation for rater reliability}
\usage{
icc(
  data,
  method = c("oneway", "agreement", "consistency"),
  format = "wide",
  cols = colnames(data),
  levels = c("id", "rater"),
  sem = TRUE,
  confint = TRUE,
  alpha = 0.05,
  var = FALSE
)
}
\arguments{
\item{data}{data.frame with repeated measures or observations in the columns
and rated subjects in the rows}

\item{method}{type of ICC that is returned, options are:
`c("oneway", "agreement", "consistency")`, the default returns all.}

\item{format}{character string indicating the data structure as either "wide"
or "long". Default `format = "wide"`.}

\item{cols}{column names used for the repeated measures in the wide format,
default uses `cols = colnames(data)`}

\item{levels}{character vector for the column names used for the levels.
Default is `levels = c("id", "rater")`.}

\item{sem}{logical vector if standard error of measurement is returned.}

\item{confint}{logical vector if confidence interval for ICC is returned.}

\item{alpha}{the confidence level required, default `alpha = 0.05`.}

\item{var}{logical vector if variance estimates are returned.}
}
\value{
matrix with relevant output
}
\description{
Intra class correlation for rater reliability
}
\examples{
mam <-
breast[,c("Mam1_score","Mam2_score", "Mam3_score")]
icc(data = mam, confint = TRUE, var = TRUE)
pch <- breast[,c("PCH1_score", "PCH2_score","PCH3_score",
"PCH4_score","PCH5_score")]
icc(data = pch)
icc(data = pch, confint = FALSE, var = TRUE)
}
\seealso{
[icc_agreement()] [icc_oneway()] [icc_consistency()]
}
