% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icc_consistency.R
\name{icc_consistency}
\alias{icc_consistency}
\title{ICC consistency}
\usage{
icc_consistency(model, alpha = 0.05)
}
\arguments{
\item{model}{merMod object result from \code{icc_model()}.}

\item{alpha}{confidence interval level, default \code{alpha = 0.05}.}
}
\value{

}
\description{
ICC consistency
}
