% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icc_model.R
\name{icc_model}
\alias{icc_model}
\title{ICC model}
\usage{
icc_model(data, cols = colnames(data))
}
\arguments{
\item{data}{data.frame with a column for each observer/rater and a row per
rated subject.}

\item{cols}{character vector with the column names to be used as observers.
Default is `cols = colnames(data)`.}
}
\value{
`lmer`model object
}
\description{
Calculate model for the ICC estimations. The basic model contains a random
intercept for the id-level and a random intercept for the observer level.
For each ICC type, different variance components are used.
}
