% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icc_model.R
\name{icc_model2}
\alias{icc_model2}
\title{ICC model}
\usage{
icc_model2(data, levels = c("id", "rater"))
}
\arguments{
\item{data}{data.frame in a long format with multiple rows per subject (id).
The data should contain a column for each of the `levels` and a column named
`score` for the scores.}

\item{levels}{character vector with the column names of the levels that should
be used in the data. Default is `cols = c("id", "rater")`.}
}
\value{
`lmer`model object
}
\description{
Calculate model for the ICC estimations. The basic model contains a random
intercept for the id-level and a random intercept for the rater level, more
levels can be indicated for a 3-way model.This fuction works from a long data
format.
For each ICC type, different variance components are used.
}
