% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icc_oneway.R
\name{icc_oneway}
\alias{icc_oneway}
\title{ICC oneway}
\usage{
icc_oneway(model, alpha = 0.05)
}
\arguments{
\item{model}{merMod object result from \code{icc_model()}.}

\item{alpha}{confidence interval level, default \code{alpha = 0.05}.}
}
\value{
a list with parameter estimates.
}
\description{
A function that computes the one-way ICC, with the corresponding standard error of measurement, sem. And the confidence intervals. see details for more information.
}
