% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_icc.R
\name{n_icc}
\alias{n_icc}
\title{Sample size for ICC}
\usage{
n_icc(beta = 0.8, alpha = 0.05, k = 3, icc, icc_lower)
}
\arguments{
\item{beta}{power level, default \code{beta = 0.8}, also called assurance
probability.}

\item{alpha}{confidence level, default \code{alpha = 0.05}}

\item{k}{number of raters, default \code{k = 3}}

\item{icc}{reliability coefficient}

\item{icc_lower}{lower limit of the reliability coefficient icc}
}
\description{
Sample size for ICC
}
\details{
Calculate the sample size for achieving a certain lower limit of the
confidence interval derived by the F procedure. This method is developed for
the ICC type oneway.
}
\examples{
n_icc(icc = 0.7, icc_lower = 0.6, k = 4)
n_icc(icc = 0.7, icc_lower = 0.6, k = 2)
n_icc(icc = 0.7, icc_lower = 0.6, k = 3)
}
\references{
Zou, G.Y. (2012) Sample size formulas for estimating intraclass correlation
coefficients with precision and assurance. Statistics in medicine, 31, 3971-3981.
}
